aanpassing for sqlserver

+++ sql/V0041__add_department.sql
@@ -1,6 +1,6 @@
 create table HR.DEPARTMENTS
 (
-  department_id   NUMBER(4) not null,
+  department_id   int not null,
   department_name VARCHAR2(30),
   manager_id      NUMBER(6),
   location_id     NUMBER(4)
